requirements.txt
dash==2.18.2
dash_ag_grid==31.3.0
dash_bootstrap_components==1.7.1
msal==1.31.1
numpy==1.23.5
pandas==1.5.3
plotly==5.9.0
Requests==2.32.3
scipy==1.10.0
shiny==1.2.1
shinywidgets==0.4.2
websocket_client==0.58.0
